requirements.txt
Flask
Flask-SocketIO
psycopg2-binary